XML parser refactored
